requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
blobfile==1.2.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
clip @ git+https://github.com/openai/CLIP.git@8cad3a736a833bc4c9b4dd34ef12b52ec0e68856
DALL-E==0.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
ftfy==6.0.3
idna==3.2
importlib-metadata==3.10.1
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
olefile==0.46
opencv-python==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655818400/work
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.3
requests==2.26.0
Send2Trash==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
zipp==3.5.0
